ms2/spaces view (#423)
* refactor(ms2): my environments -> my spaces

* (ms2/my-spaces): removed delete button

* chore(ms2/my-spaces): removed side panel

* chore(ms2/my-spaces): removed selection

* feat(ms2/my-spaces): enter space button

* refactor(ms2/my-spaces): use next link

* typo in filename

---------
