+++ src/management-system-v2/app/(dashboard)/[environmentId]/layout-client.tsx
@@ -96,9 +96,9 @@ const Layout: FC<
               icon: <FaUserEdit />,
             },
             {
-              key: 'environments',
+              key: 'spaces',
               title: 'My Spaces',
-              label: <SpaceLink href={`/environments`}>My Spaces</SpaceLink>,
+              label: <SpaceLink href={`/spaces`}>My Spaces</SpaceLink>,
               icon: <AppstoreOutlined />,
             },
           ],
@@ -187,7 +187,6 @@ const Layout: FC<
                     priority
                   />
                 </Link>
-
                 {loggedIn ? menu : null}
               </AntLayout.Sider>
             )}
